Move getPid to helpers file
The extras_containers.go file is only built on amd64. This prevents
the use of this function by the cgroup parser extension on non-amd64
architectures.
